NodeDetailのロジックをカスタムフックに切り出すリファクタ
